% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{col2hex}
\alias{col2hex}
\title{Convert colors to hex RGB}
\usage{
col2hex(x)
}
\arguments{
\item{x}{A single or multiple color names or numeric values.}
}
\value{
A single or multiple hexadecimal color strings.
}
\description{
This function accepts one or more colors, provided as integer numeric values
or character values, and converts them into hexadecimal RGB colors, such as
"#00FF00".
}
\author{
Johannes B. Gruber, Philip Leifeld
}
