% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{dna_addAttribute}
\alias{dna_addAttribute}
\title{Add an attribute to the DNA database}
\usage{
dna_addAttribute(
  connection,
  statementType = 1,
  variable = "organization",
  value = "new value",
  color = "#000000",
  type = "",
  alias = "",
  notes = "",
  returnID = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{connection}{A \code{dna_connection} object created by the
\code{dna_connection} function.}

\item{statementType}{The ID of the statement type (as an integer) or the name
of the statement type (as a character object) in which the variable is
defined.}

\item{variable}{The name of the variable for which attribute data should be
retrieved, for example \code{"organization"} or \code{"concept"}.}

\item{value}{The value of the new attribute as a character object.}

\item{color}{A character object containing the color of the new document as a
hexadecimal RGB value.}

\item{type}{The type of the new attribute as a character object.}

\item{alias}{A character object containing the alias of the new attribute.}

\item{notes}{The notes of the new attribute as a character object.}

\item{returnID}{Return the ID of the newly created attribute as a numeric
value?}

\item{verbose}{Print details?}
}
\description{
Add a new attribute to the DNA database.
}
\details{
The \code{dna_addAttribute} function can add new attributes to an existing
DNA database. Attributes are the annotations that are coded within a
variable, along with some meta-data. For example, if the variable is
"organization", then an attribute value could be "Environmental Protection
Agency", and other attribute meta-data for this value could be its color,
actor type etc. The \code{dna_addAttribute} function can not only be used to
add entries to the attributes table in an ongoing project, but also to
pre-populate a DNA database with values for deductive coding. Either way, the
user supplies a \link{dna_connection} object as well as various details about
the attribute to be added, for example the value, color, type etc. The
attribute ID will be automatically generated and can be returned if
\code{returnID} is set to \code{TRUE}.
}
\author{
Philip Leifeld
}
