% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{dna_getDocuments}
\alias{dna_getDocuments}
\title{Retrieve a dataframe with documents from a DNA connection}
\usage{
dna_getDocuments(connection)
}
\arguments{
\item{connection}{A \code{dna_connection} object created by the
\code{dna_connection} function.}
}
\description{
Retrieve a dataframe with all documents from a DNA connection.
}
\details{
This function creates a dataframe with one row per document and contains
columns for the document ID, title, the complete text, and all meta data. The
dataframe can then be manually manipulated and returned to the DNA database
through the \link{dna_setDocuments} function.
}
\examples{
\dontrun{
dna_init()
conn <- dna_connection(dna_sample())
documents <- dna_getDocuments(conn)
documents$title[1] <- "New title for first document"
documents$notes[3] <- "Added a note via rDNA."
documents <- documents[, -5]  # Removing the fifth document
dna_setDocuments(conn, documents, simulate = TRUE)  # apply changes
}

}
\author{
Philip Leifeld
}
