% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{dna_getSettings}
\alias{dna_getSettings}
\title{Retrieve a dataframe with all settings}
\usage{
dna_getSettings(connection)
}
\arguments{
\item{connection}{A \code{dna_connection} object created by the
\code{dna_connection} function.}
}
\description{
Retrieve a dataframe with all settings related to the GUI.
}
\details{
The \code{dna_getSettings} function retrieves all settings related to the
graphical user interface, including pop-up window width, active coder etc.
}
\author{
Philip Leifeld
}
