% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{dna_installJar}
\alias{dna_installJar}
\title{Install DNA JAR file into \code{inst/java/} sub-directory}
\usage{
dna_installJar()
}
\value{
The function returns \code{0} upon success and \code{1} upon failure.
}
\description{
Install DNA JAR file into \code{inst/java/} sub-directory of current path.
}
\details{
rDNA requires the installation of a DNA JAR file to run properly. While it is
possible to store the JAR file in the respective working directory, it is
preferable to install it in the rDNA package installation directory under
\code{inst/java/}. The \code{dna_installJar} function attempts doing so. It
is called automatically during rDNA package installation, but it can also be
called manually. If it is called manually by the user, the working directory
must be the root directory of the source files of the rDNA package (i.e., the
directory where the \code{DESCRIPTION} file is located).

The function first attempts to locate a JAR file in the \code{inst/java/}
directory. If a JAR file is found there, the function exits, and installation
of the package can normally resume, with the JAR file being installed in the
\code{java} sub-directory of the installation directory. If no JAR file can
be located, an attempt is made to download a JAR file with the same version
number as the rDNA package in the file name from the GitHub release page of
the DNA repository. The JAR file is then stored under \code{inst/java/} and
installed in the rDNA installation directory. If this also fails, the
function then attempts to download the most recent source code of DNA from
the \code{master} branch of DNA on GitHub and compile the DNA JAR file from
source. This can only work if the local system has at least version 8.0 of
the Java Development Toolkit (JDK) installed. The JDK can be downloaded from
\url{https://jdk.java.net/}. Open JDK versions are permitted as an
alternative. The JAR file is only compiled from source and used for
installation if the latest DNA version on the GitHub master branch has the
same version number as the rDNA package. If this also fails, the function
exits and prints instructions on downloading and installing the DNA JAR file
manually.
}
\seealso{
\code{\link{dna_installJar}}
}
\author{
Philip Leifeld
}
