% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{dna_removeAttribute}
\alias{dna_removeAttribute}
\title{Removes an attribute entry from the database}
\usage{
dna_removeAttribute(
  connection,
  statementType = 1,
  variable = "organization",
  attribute,
  removeStatements = FALSE,
  simulate = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{connection}{A \code{dna_connection} object created by the
\link{dna_connection} function.}

\item{statementType}{The ID of the statement type (as an integer) or the name
of the statement type (as a character object) in which the variable is
defined.}

\item{variable}{The name of the variable from which attribute should be
removed, for example \code{"organization"} or \code{"concept"}.}

\item{attribute}{An integer value denoting the ID of the attribute to be
removed, or a value label (as a character object) denoting the entry to be
removed. The \link{dna_getAttributes} function can be used to look up IDs.}

\item{removeStatements}{The attribute given by \code{attribute} may contain
statements. If \code{removeStatements = TRUE} is set, these statements
are removed along with the respective attribute. If
\code{removeStatements = FALSE} is set, the statements are not deleted,
the attribute is kept as well, and a message is printed.}

\item{simulate}{Should the changes only be simulated instead of actually
applied to the DNA connection and the SQL database? This can help to
plan more complex recode operations.}

\item{verbose}{Print details on whether the attribute could be removed?}
}
\description{
Removes an attribute entry from the database based on its ID or value.
}
\details{
The user provides a connection object and the ID or value label of an
existing attribute in the DNA database, and this attribute is removed both
from memory and from the SQL database, possibly including any statements
associated with the attribute.
}
\author{
Philip Leifeld
}
