% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{dna_removeRegex}
\alias{dna_removeRegex}
\title{Removes a statement from the database}
\usage{
dna_removeRegex(connection, regex)
}
\arguments{
\item{connection}{A \code{dna_connection} object created by the
\link{dna_connection} function.}

\item{regex}{The regular expression that should be removed as a character
object.}
}
\description{
Removes a statement from the database based on its ID.
}
\details{
The user provides a connection object and the ID of an existing statement in
the DNA database, and this statement is removed both from memory and from the
SQL database.
}
\author{
Philip Leifeld
}
