% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{print.dna_scale}
\alias{print.dna_scale}
\title{Print the summary of a \code{dna_scale} object}
\usage{
\method{print}{dna_scale}(x, ...)
}
\arguments{
\item{x}{A \code{dna_scale} object.}

\item{...}{Further options (currently not used).}
}
\description{
Show details of the MCMC chain created by the four \code{dna_scale}
functions.
}
\details{
Prints the method that was used for the scaling, the number of values in the
object and their means.
}
\examples{
\dontrun{
dna_init()
conn <- dna_connection(dna_sample())
dna_scale <- dna_scale1dbin(conn,
                            variable1 = "organization",
                            variable2 = "concept",
                            qualifier = "agreement",
                            threshold = 0.51,
                            mcmc_iterations = 20000,
                            mcmc_burnin = 2000,
                            mcmc_thin = 10,
                            store_variables = "both")
dna_scale
}
}
\author{
Tim Henrichsen, Johannes B. Gruber
}
